Split the IO module into base and high level IO
