+++ tests/tests.c
@@ -1,6 +1,6 @@
 #include "unity.h"
 #include "data.h"
-#include "io.h"
+#include "high-io.h"
 #include "logger.h"
 #include "memory.h"
 #include "numeric.h"
