+++ walls-of-doom/about.c
@@ -1,7 +1,8 @@
 #include "about.h"
 #include "constants.h"
 #include "data.h"
-#include "io.h"
+#include "logger.h"
+#include "high-io.h"
 #include <string.h>
 
 /**
