+++ walls-of-doom/main.c
@@ -1,4 +1,4 @@
-#include "io.h"
+#include "high-io.h"
 #include "logger.h"
 #include "memory.h"
 #include "menu.h"
