+++ walls-of-doom/record.c
@@ -2,7 +2,7 @@
 #include "code.h"
 #include "constants.h"
 #include "data.h"
-#include "io.h"
+#include "high-io.h"
 #include "logger.h"
 #include "memory.h"
 #include "numeric.h"
